requirements.txt
absl-py==0.1.11
astor==0.6.2
bleach==1.5.0
cycler==0.10.0
decorator==4.2.1
gast==0.2.0
grpcio==1.10.0
h5py==2.7.1
html5lib==0.9999999
joblib==0.11
Keras==2.1.5
Markdown==2.6.11
matplotlib==2.1.2
networkx==2.1
numpy==1.14.2
opencv-python==3.4.0.12
Pillow==5.0.0
protobuf==3.5.2.post1
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2018.3
PyWavelets==0.5.2
PyYAML==3.12
scikit-image==0.13.1
scipy==1.0.0
six==1.11.0
tensorboard==1.6.0
tensorflow==1.6.0
tensorflow-gpu==1.6.0
termcolor==1.1.0
Theano==1.0.1+53.g546067d
tqdm==4.19.8
Werkzeug==0.14.1
